feat(tool): make paint and getColor utils functions
